% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_names_from_text.R
\name{bird_names_from_text}
\alias{bird_names_from_text}
\title{bird names from text}
\usage{
bird_names_from_text(x)
}
\arguments{
\item{x}{one or more character strings, e.g. a database notes field. Will recognize alpha codes and common names.}
}
\value{
A character vector of same length as x. If multiple bird species names exist in
each element of x, they will be separated by an underscore in the output vector.
}
\description{
Extract bird species names from text.
}
\examples{
test_text <- c("Bald Eagle and White-tailed Kite and WREN are birds.",
"The Great Blue Heron is taller than the Snowy Egret")

bird_names_from_text(test_text)
}
