% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_bbl}
\alias{download_bbl}
\title{Download BBL species list
bbl list has taxonomic numbers, which AOU discontinued in the 7th checklist update.
current version of birdnames does taxonomic ordering based on the AOU list.
bbl list probably not needed but retaining this function for possible future use of bbl list
(e.g. band size or T&E status)}
\usage{
download_bbl()
}
\value{
data frame
}
\description{
Download BBL species list
bbl list has taxonomic numbers, which AOU discontinued in the 7th checklist update.
current version of birdnames does taxonomic ordering based on the AOU list.
bbl list probably not needed but retaining this function for possible future use of bbl list
(e.g. band size or T&E status)
}
\examples{
# downloads take a while so commented out for package checking
# download_bbl
}
