% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_genus}
\alias{extract_genus}
\title{Extract genus from a scientific name}
\usage{
extract_genus(df)
}
\arguments{
\item{df}{Data frame with species field. Will most likely be birdpop_df.}
}
\value{
Data frame
}
\description{
Helper function for merging birdpop and aou lists; not exported.
}
\examples{
# Will generally be piped together to create bird_list

}
