% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_taxonomy}
\alias{fill_taxonomy}
\title{Fill in higher taxonomic information
birdpop_df includes some records for birds ID to higher level than genus, but separate_species()}
\usage{
fill_taxonomy(df)
}
\arguments{
\item{df}{data frame output from separate_species()}
}
\value{
Data frame
}
\description{
Fill in higher taxonomic information
birdpop_df includes some records for birds ID to higher level than genus, but separate_species()
}
\examples{
# Will generally be piped together to create bird_list

}
