% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_custom_bird_list.R
\name{make_custom_bird_list}
\alias{make_custom_bird_list}
\title{Compile custom bird list}
\usage{
make_custom_bird_list()
}
\value{
data frame matching structure of full_bird_list
}
\description{
Helper function to combine bird group dataset and user-filled dataset.
}
\examples{
custom_bird_list <- make_custom_bird_list()
}
