% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_full_bird_list.R
\name{make_full_bird_list}
\alias{make_full_bird_list}
\title{Make full bird species list}
\usage{
make_full_bird_list(from = c("web"), read_from = NA)
}
\arguments{
\item{from}{use "web" to read tables direct from websites,
use "disk" to read from location saved to disk}

\item{read_from}{if from = "disk", specify location where csv's are saved,  as birdpop_list and aou_list}
}
\value{
data frame with columns: common.name, alpha.code, species, order, family, subfamily, genus
}
\description{
Make a list of North American bird species, combining:
4-letter (banding) codes from https://www.birdpop.org/docs/misc/Alpha_codes_tax.pdf
higher taxonomic classifications from http://checklist.aou.org/taxa
numeric species codes from https://www.pwrc.usgs.gov/bbl/manual/speclist.cfm
}
\examples{
# full_bird_list <- make_full_bird_list(from = "web")

}
