% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scrape_bbl_species_list}
\alias{scrape_bbl_species_list}
\title{Title}
\usage{
scrape_bbl_species_list()
}
\value{

}
\description{
Title
}
\examples{
bbl_list <- scrape_bbl_species_list()
saveRDS(bbl_list, "C:/Users/scott.jennings/Documents/Projects/birdnames_support/data/bbl_list")
}
