% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{separate_species}
\alias{separate_species}
\title{separate species field}
\usage{
separate_species(df)
}
\arguments{
\item{df}{Data frame with species field. Will most likely be output from extract_genus() \%>\%
left_join(aou \%>\% dplyr::distinct(genus, subfamily, family, order))}
}
\value{
Data frame
}
\description{
Separate species field into genus (dropped), specific epithet ("species.name") and subspecies name, if applicable.
}
\examples{
# Will generally be piped together to create bird_list

}
