% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_alpha.R
\name{update_alpha}
\alias{update_alpha}
\title{Update old alpha codes}
\usage{
update_alpha(alpha.code)
}
\arguments{
\item{alpha.code}{character}
}
\value{
character
}
\description{
Update old alpha codes to reflect bird name changes.
}
\examples{
update_alpha("OLDS")
update_alpha(c("OLDS", "MALL", "BRAN"))

}
