% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_birdpop}
\alias{download_birdpop}
\title{Download species list from birdpop.org}
\usage{
download_birdpop()
}
\value{
a data frame
}
\description{
The birdpop.org list has alphabetical codes (banding codes) for species, subspecies and some higher taxonomic groups
}
\examples{
# downloads take a while so commented out for package checking
# download_birdpop()
}
